+++ tests/selection-manipulation.rs
@@ -1,6 +1,7 @@
 mod data;
 
 use data::doc_with_siblings;
+use dom_query::Document;
 
 #[test]
 fn test_replace_with_html() {
@@ -60,4 +61,29 @@ fn test_replace_with_selection() {
     assert!(sel.is("#nf6"));
     assert_eq!(doc.select("#nf6").length(), 0);
     assert_eq!(doc.select("#nf5").length(), 1);
+}
+
+#[test]
+fn test_create_element() {
+    let contents = r#"<!DOCTYPE html>
+    <html lang="en">
+        <head></head>
+        <body>
+            <div id="main">
+            <div>
+        </body>
+    </html>"#;
+
+    let doc = Document::from(contents);
+
+    let el = doc.tree.new_element("p");
+    el.set_attr("id", "inline");
+
+    let main_id = doc.select_single("#main").nodes().iter().next().unwrap().id;
+
+    doc.tree.append_child_of(&main_id, &el.id);
+
+    //TODO: easy way to get parent id? or an easier way to append a child node without id?
+
+    assert!(doc.select("#main #inline").exists());
 }